+++ src/services/public-book.js
@@ -3,7 +3,7 @@ import shttp from 'utils/shttp';
 export default {
   async getBookInfo(params) {
     const result = await shttp({
-      url: `/v1/book/${params.id}`
+      url: `/v1/public/book/${params.id}`
     });
     return { item: result.data };
   },
@@ -13,13 +13,13 @@ export default {
       search += `&${k}=${params.query[k]}`
     }
     const result = await shttp({
-      url: `/v1/book/${params.id}/catalog?${search}`
+      url: `/v1/public/book/${params.id}/catalog?${search}`
     });
     return { items: result.data, ended: result.data.length < 10 };
   },
   async getBookChapter(params) {
     const result = await shttp({
-      url: `/v1/book/${params.bid}/chapter/${params.id}`
+      url: `/v1/public/book/${params.bid}/chapter/${params.id}`
     });
     return { item: result.data };
   },