+++ CHANGELOG.md
@@ -28,3 +28,8 @@
 - VisuaBoxView/SwitchView/AutoCenterView
 - mock-server添加控制配置
 - isDebug和console配置
+
+## mater 2019-10-3 03:01:08
+- web-api-admin三端,public-user-admin三角色
+- 真实接口联调
+- state传值